+++ Rubrik/Private/Get-RubrikAPIData.ps1
@@ -1723,7 +1723,21 @@ function Get-RubrikAPIData($endpoint) {
                 Filter      = ''
                 Success     = '202'
             }
-        }        
+        }
+        'Update-RubrikVMwareVM'         = @{
+            '1.0' = @{
+                Description = 'Refresh the metadata for the specified VMware VM'
+                URI         = '/api/internal/vmware/vcenter/{id}/refresh_vm'
+                Method      = 'Post'
+                Body        = @{
+                    vmMoid = 'vmMoid'
+                }
+                Query       = ''
+                Result      = ''
+                Filter      = ''
+                Success     = '204'
+            }
+        }      
 
     } # End of API